requirements.txt
cairo_lang==0.7.0
cryptowatch==0.0.11
cryptowatch_sdk==0.0.16
pytest==6.2.5
python-decouple==3.6
requests==2.22.0
SQLAlchemy==1.3.12
starknet.py==0.1.6a0
starknet_py==0.1.8a0
